notes fade out, combined play/stop button
add MyMediaPlayer that plays a mediaplayer file for a given amount of
time and then gradually fades out the audio so that it doesn't cut off
abruptly. Made the play and stop button  into one button. When nothing
is being played the play sign is shown, when something is being played
then the stop sign is shown.